requirements.txt
pandas
openai
python-dotenv
sentence-transformers
torch